requirements.txt
astroid==2.0.4
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
defusedxml==0.5.0
Django==2.1.2
django-bootstrap4==0.0.7
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.8.2
drf-writable-nested==0.5.1
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
psycopg2==2.7.5
PyJWT==1.6.4
pylint==2.1.1
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
typed-ast==1.1.0
urllib3==1.24.1
wrapt==1.10.11
